Module moore_svlog_syntax::ast

Re-exports

pub use self::ExprData::*;

pub use self::StmtData::*;

pub use self::TypeData::*;

Structs

Assertion

CallArg

ClassDecl

ClassItem

Constraint

ConstraintItem

ContAssign
A continuous assignment statement.

CycleDelay

DelayControl

EnumName

EventControl

Expr

GenerateBlock
A body of a generate construct. May contains hierarchy items or more generate constructs.

GenerateCase

GenerateFor

GenerateIf

GenvarDecl

Identifier

ImportDecl

ImportItem

Inst

InstName

IntfDecl

ModDecl

ModportDecl

ModportItem

NetDecl

NodeId
A positive, small ID assigned to each node in the AST. Used as a lightweight way to refer to individual nodes, e.g. during symbol table construction and name resolution.

PackageDecl

ParamAssignment

ParamDecl
A parameter or localparam declaration.

ParamTypeDecl
A single type assignment within a parameter or localparam declaration.

ParamValueDecl
A single value assignment within a parameter or loclparam declaration.

PatternField

PortConn
A port connection as given in an instantiation.

PortDecl

Procedure

PropExpr

PropSpec

Root

SeqExpr

Stmt

StreamExpr

StructMember

SubroutineDecl

SubroutinePort

SubroutinePortDecl

SubroutinePortName

SubroutinePrototype

Timeunit

Type

Typedef

VarDecl

VarDeclName

Enums

AssertionActionBlock

AssertionData

AssignOp

BlockingAssertion

CaseItem

CaseKind

CaseMode

ChargeStrength

ClassItemData

ClassItemQualifier

ConcurrentAssertion

ConstraintItemData

ConstraintKind

DriveStrength

EdgeIdent

EventControlData

EventExpr

ExprData

HierarchyItem

Item

JoinKind

Lifetime
Lifetime specifier for variables, tasks, and functions. Defaults to static.

ModportPort

NetKind

NetStrength

NetType

ParamKind

PatternFieldData

Port

PortConnKind

PortConnMode
Represents how a named port connection is made.

PortDir

PortKind

ProcedureKind

PropBinOp

PropExprData

PropSeqBinOp

PropSeqOp

RandomQualifier

RangeMode

SeqBinOp

SeqExprData

SeqRep

StmtData

StreamConcatSlice

StructKind

SubroutineItem

SubroutineKind

SubroutinePortDir

TimingControl

TypeData

TypeDim

TypeOrExpr

TypeSign

UniquePriority

Constants

DUMMY_NODE_ID
During parsing and syntax tree construction, we assign each node this ID. Only later, during the renumbering pass do we assign actual IDs to each node.